fixed merge conflicts.
The main change is #302 removed once_cell in favor of
std::sync::OnceLock, based on some git history, it seems like we had
some performance issues with having buffer pool as a global pool for
short lived connections (like PHP for example). We have made some
changes so the buffer pool is local to the connection.